FogBugz #368472: Support SDMMC in Stratix10 SWVP device tree.
This patch is for socfpga-4.3 kernel version only.  Version 1.1 of
the Stratix10 SWVP is slated to use kernel version 4.3, and this
patch is needed for code-completion of S10 SWVP 1.1

arm64: dts: Support SDMMC on Stratix10 SWVP

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>
---
V2: Use decimal for frequency values instead of hex.
V3: Use 50MHz instead of 25MHz.